backend/requirements.txt
Django>=4.2.27
djangorestframework
django-cors-headers
psycopg2-binary
python-socketio[asyncio]
uvicorn[standard]
PyJWT
requests